% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMTL.R
\name{plot_HMTL}
\alias{plot_HMTL}
\title{Plot diagram of the information criterion vs. penalty parameters}
\usage{
plot_HMTL(x)
}
\arguments{
\item{x}{Object of class "Model Selection" created by \code{Selection_HMTL}}
}
\value{
X axis represents the value of penalty parameters, and y axis represents the estimated values of the composite likelihood, degree of freedom for model complexity, and the (robust) Bayesian information criterion.
}
\description{
Plot a diagram to illustrate the change of Bayesian information criterion with different penalty paameters for model selection
}
